# backend/historial-service/swagger.yaml
openapi: 3.0.0
info:
  title: Historial API
  description: API para gestionar el historial de videos vistos.
  version: 1.0.0
paths:
  /historial:
    post:
      summary: Agregar un video al historial
      operationId: addToHistorial
      requestBody:
        description: Video para agregar al historial
        content:
          application/json:
            schema:
              type: object
              properties:
                video_id:
                  type: integer
                title:
                  type: string
                genre:
                  type: string
      responses:
        '200':
          description: Video agregado al historial
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  video:
                    type: object
                    properties:
                      video_id:
                        type: integer
                      title:
                        type: string
                      genre:
                        type: string

    get:
      summary: Obtener el historial de videos
      operationId: getHistorial
      responses:
        '200':
          description: Lista de videos en el historial
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    video_id:
                      type: integer
                    title:
                      type: string
                    genre:
                      type: string

    delete:
      summary: Eliminar todos los videos del historial
      operationId: deleteHistorial
      responses:
        '200':
          description: Historial de videos eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Historial eliminado exitosamente"

  /historial/{video_id}:
    put:
      summary: Eliminar un video específico del historial
      operationId: removeVideoFromHistorial
      parameters:
        - name: video_id
          in: path
          description: ID del video que se quiere eliminar del historial
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Video eliminado del historial
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Video eliminado exitosamente"